requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-openai
openai
faiss-cpu
python-dotenv
fastapi
uvicorn
python-multipart